fix: template error
in Row::get_data_at, I was returning a reference inside a std::expected,
which is not allowed. Switching to std::reference_wrapper solves the
issue though.

In the Row::set_data_at function, the std::get template should be
Column<T>, not just T, because the member m_cols contains variants of
Columns.

In RowBuilder::set_rowid, the check is_rowid_set should be checking if
true, not checking if false.

commit-by: Huy Nguyen