feat: new ColumnBuilder, start making Table
Move Column's constructors to private, only allowing friend classes (in
this case, ColumnBuilder) to construct Column. This will let me expose
the Column class to the API, but since I don't need to expose
ColumnBuilder to the public API, the end-user cannot construct columns
by themselves.

I plan for all column constructions to be handled by the DbFile or at
least the Table.

Change column.hxx and table.hxx to use the aliased data type.

Change main() to use the column builder. Of course, it's still
placeholder, and will be removed one day.

commit-by: Huy Nguyen