column flags, templated column builder.
Added a new enum class called ColumnFlags. In the future this can also
mark columns as secondary key, foreign key, nullable, and so on.

Also defined some convenient operator and functions to operate on
ColumnFlags. I bet most of us will choose the OR operator to chain
stuff, but I also provide a concrete function `combine`, which does the
exact same thing.

ColumnBuilder is now converted into a template. Since now I need more
configuration (aka, the builder needs to save the column whilst building
it), and Column is templated, ColumnBuilder is naturally templated.

The main() function is currently holding tests for my ColumnFlags. Took
me quite a while to figure out the correct way to deal with the
parameter pack of `combine`, and also I realized that this languages
doesn't know how to print out `std::uint8_t`. I guess it's because it
hallucinates this with `unsigned char`.

commit-by: Huy Nguyen