feat: add a method to quickly build a column
Rather than however I have to call that ColumnBuilder method. Getting it
nested inside a RowBuilder also, makes it an absolute nightmare.

Anyways, I should design an API that helps the user NOT to think about
columns and rows individually. Because rows and columns that don't
belong to a table is kinda worthless.

commit-by: Huy Nguyen