fix: ModuleIfEnabled with dummy file
I forgot that a non-interface library always need a source file. But,
for a module-only library, which does nothing in no-module builds, there
can be no source files. So, I resort to an empty source file inside the
`cmake` directory called `dummy.cxx`. It was there before IIRC. But I
thought I wouldn't need it, so I deleted it.

commit-by: Huy Nguyen