start writing TableBuilder
The TableBuilder needs to be separate because I don't want to expose the
Table constructor to the end user. They should only build a table in the
context of a database file (which, I haven't written yet, but soon
enough). And if they are doing like so, it's the database file builder
that exposes this functionality to them.

commit-by: Huy Nguyen